fixing bug whereby local storage get's cleared nearly everywhere

+++ src/js/contentScript.js
@@ -1,4 +1,4 @@
-if (!matchDomain(['seekingalpha.com', 'sfchronicle.com', 'cen.acs.org'])) {
+if (isSiteEnabled({ url: window.location.href }) && !matchDomain(['seekingalpha.com', 'sfchronicle.com', 'cen.acs.org'])) {
   window.localStorage.clear();
 }
